Fix max album cover height.  Move player layout settings to sub menu.
